fix Bill

+++ HospitalManagementSystem/ViewModels/AppointmentPatientDetailsViewModel.cs
@@ -1,6 +1,8 @@
 ﻿using HospitalManagementSystem.Models;
 using HospitalManagementSystem.Services;
 using HospitalManagementSystem.Views;
+using HospitalManagementSystem.Views.Components;
+using MaterialDesignThemes.Wpf;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
@@ -21,7 +23,7 @@ namespace HospitalManagementSystem.ViewModels
         public String PatientBirthDate { get; set; }
         public String PatientDiagnosis { get; set; }
         public String PatientBill { get; set; }
-
+        public String textValidation { get; set; }
         // Edit
         public ICommand editAppointmentPatient { get; set; }
         public ICommand deleteAppointmentPatient { get; set; }
@@ -39,6 +41,7 @@ namespace HospitalManagementSystem.ViewModels
 
         public AppointmentPatientDetailsViewModel(String id)
         {
+            PatientBill = ((AppointmentPatient)(Hospital.Patients[id])).getBill().ToString("0.00") + '$';
             PatientID = id;
             DoctorsList = new ObservableCollection<ComboBoxPairs>();
             AppointmentsList = new ObservableCollection<ComboBoxPairs>();
@@ -47,7 +50,6 @@ namespace HospitalManagementSystem.ViewModels
                 DoctorsList.Add( new ComboBoxPairs(doctor.ID, doctor.Name));
             }
             DoctorsNumber = "Doctors: " + Hospital.Patients[id].Doctors.Count.ToString();
-
             foreach(Appointment appointment in ((AppointmentPatient)Hospital.Patients[id]).Appointments.Values)
             {
                 AppointmentsList.Add(new ComboBoxPairs(appointment.ID, ("Date: " + appointment.Date.ToString() + " | " +appointment.Doctor.Name)));
@@ -75,12 +77,16 @@ namespace HospitalManagementSystem.ViewModels
 
         }
 
-        public void DeleteAppointmentPatient()
+        public async void DeleteAppointmentPatient()
         {
+            object result = await DialogHost.Show(new DeleteMessageBox(), "RootDialog");
+            if (result.Equals(true))
+            {
+                // Delete Logic Here
             HospitalDB.DeletePatient(PatientID);
             Hospital.DeletePatient(PatientID);
-            Home.ViewModel.CloseRootDialog();
-            Home.ViewModel.Content = new PatientsViewModel();
+            Home.ViewModel.Content = new PatientsViewModel();            
+            }
         }
     }
 }
